prevent crash when searching contacts
We add a check in the GtkQTreeModel api to make sure that the
GtkTreeIter passed is a valid one for the given model.

For some reason, it seems that in the ContactsView, the GtkTreeView
sometimes gives us an invalid GtkTreeIter in the GtkTreeCellDataFunc
callback (eg: render_contact_photo()). This patch prevents the
client from crashing due to a bad QModelIndex. However, it would
be worth invesgitating why we're receving an invalid GtkTreeIter
in the first place.

Change-Id: I0c00f22ea5adbf877b73e2cedb7eacbc20b3f3b6
Tuleap: #1448