GtkQTreeModel: fix layoutChanged handling
The layoutChanged signal was not being handled correctly. It is
possible for items to be added or removed during a layoutChanged,
so simply emiting a row-changed signal is not enough. Since GTK+
has no equivalend to layoutChanged in the GtkTreeModel interface,
we simply remove and then re-add all of the rows. This gurantees
that our view remains synchronized with the Qt model.

Change-Id: I8c7f1ac0a608a9f3fd386878ff50e10b2ed655c2