Update patch set 14
Patch Set 14: Code-Review-1

So the PixmapManipulatorI interface (pixbufmanipulator for gnome) takes dislayPresence as an argument. See the callPhoto() and contactPhoto() functions. I think instead of checking the presence in the recentcontactsview, you should move the check to those functions in case the called sets the displayPresence parameter to true, and then call your new drawing functions in there.

Patch-set: 14
Label: Code-Review=-1