prevent crash during Qt model layout change
Now that we handle the layout change by removing and then
re-inserting all of the rows in the GTK model wrapper, we need to
be careful how we handle the selection changes in the tree view.
The selection will be lost during the handling of the layout change
so we must ignore this.

This patch adds a way to check if we're in the middle of a layout
change. Then we use this in the handlers to the selection change
and other relevant signals to make sure we don't lose the selection
during this.

This prevents us destroying the call / chat view prematurely while
sending a text message which caused a crash in some cases.

Change-Id: Ic46a8d9cda718c4c8058918067593ed1734f471b
Reviewed-by: Nicolas Jäger <nicolas.jager@savoirfairelinux.com>