Display name and number in the smart view
Previously we were only showing the name and if there is none then
the number. To improve usablility and security (so that users are
more aware of who they are actually communicating with) we display
both. The number we display is either the Ring registered name or
else just the URI.

This is implemented by now displaying the name in bold in the first
row of the middle column of the treeview and the number in smaller
text bellow it. The last used time/date and the call status is
moved to the last (3rd) column, to the first row. The unread
message count or the call duration will then be shown in the 2nd
row of the last column.

Since there is now less room for the last used time/date, we now
only show the time if the last interaction was "today"; otherwise
we show the date category.

Change-Id: Icb1011d88c607c0e97a779a72eca4413fbe88af4