Display name and number in the smart view
Previously we were only showing the name and if there is none then
the number. To improve usablility and security (so that users are
more aware of who they are actually communicating with) we display
both. The number we display is either the Ring registered name or
else just the URI.

Change-Id: Icb1011d88c607c0e97a779a72eca4413fbe88af4