Prevent warnings about size allocation
The overlay controls of the CurrentCallView were being placed in
the VideoWidget before the view that it is in was being allocated
its full size. This resulted in several warnings being printed
about size allocation each time a CurrentCallView is displayed.

We prevent this by not inserting the controls until we get a
size-allocate signal.

Change-Id: I1c1a42a873182a125722879ee09c037120c9611d
Tuleap: #1525
Reviewed-by: Nicolas Jäger <nicolas.jager@savoirfairelinux.com>