[WIP/idea]Add a search box for messages
It's just a quick draft, the code is dirty but it works.
In the gnome-client if we want to search a previous message, we can't. So, I
created a simple window to search messages. I don't know if you think it's a good idea,
but I can work to improve this PR.

Change-Id: Ic0da9abf8e9d579ec776ea6a51fce5de10043ba2