Update patch set 33
Patch Set 33: Code-Review-2

this patch is hard to review, but I've seen some erreurs, like a if() missplaced or with wrong successive action. But it's hard to find, du to the numerous warnings during the build hidding real problems.
Other examples are some non-longer used functions/methods, but it looks like a lost of functionality.
Solving them is a requirement before continuing the review.

Patch-set: 33
Label: Code-Review=-2