currentcallview: reactivate the quality slider
When switching to new models from the LRC, the auto-quality was disabled
because not implemented. But now, the video quality is really bad. So we
can use the old models to re-enable this functionnality for now, and: th
will replace this when we will replace the video manager.

Note, the code was taken in commit fad674f560cb9a07902b4058ba0c938299ae036c

Change-Id: Ibeb950afbe56f66a904266a2042cd9af39b2b3cb
Tuleap: #1759
Reviewed-by: Guillaume Roguez <guillaume.roguez@savoirfairelinux.com>