fix: make downloads directory dialog modal
The Download folder button (Settings/General) is implemented as a
FileChooserButton, which makes the ChooseFolder dialog non-modal by
default. While this can be meaningful in many cases, it is a very
confusing feature for Ring users.

In this patch, we:
a) Stop using FileChooserButton (it is very difficult to have modal
   dialogs with this class)

b) Instead opt for a classical button with dedicated callback handling
   dialog stuff

In order to make the dialog modal, a reference to the main window is
needed by the callback. Since the GeneralSettingsView inherits from
GtkScrolledWindow, this reference is not stored by default into the
class. While we could certainly add the reference to the class
contructor, a global variable is probably better because we are likely
to replicate these changes to a variety of other Ring classes, as many
other file choosers suffer from this bug, and a global variable would
make the general amount of changes by far smaller.

Change-Id: I8ea4fb7fa32d289abffc44ee80215f08c155cbd6
Tuleap: #1852