fix: race condition + crash when deleting accounts
Whenever an account is deleted, slotAccountRemovedFromLrc() is
called. This function retrieves the latest account list from
the LRC and (in sequence):

1) calls refreshAccountSelectorWidget to update the account
   selector

2) calls gtk_notebook_set_current_page to update the conversation
   panel

3) calls updateLrc. At this point, the accountContainer_ variable
   is updated, which means the current account has changed.

4) finally calls ring_welcome_update_view to update the main view.

During its execution, refreshAccountSelectorWidget triggers a signal
which concurrently executes the on_account_changed callback which in
turn calls onAccountSelectionChange.

onAccountSelectionChange does almost the exact same things as
slotAccountRemovedFromLrc. This is at least a waste of time,
and at most a possibly dangerous race condition. But the main issue
comes from the fact that changeView is first called to switch to the
main welcome view, and then in sequence 2), 3) and 4).

changeView calls various methods to retrieve the chats and contacts
of the current account (defined by the accountContainer_ variable).
Depending on whether updateLrc was already called by the other thread
or not, it may try to retrieve information from the deleted account,
fataly leading to various potentially security relevant crashes.

In this patch we:
* remove the duplicate code from slotAccountRemovedFromLrc
* move the updateLrc call to be before the changeView call

Change-Id: I953d9b193ce8cc3dd2a98f220f6e8b8afa3525bf