Update patch set 4
Patch Set 4: Code-Review-2

(2 comments)

The shared_ptr is a non-sens here.
The AccountContainer instance created by RingMainWindow is never modified during the life-time of the whole application with this patch.
The patch is not correct for me as it brings more complexity with a non-needed smart-pointer.
But you have the half of the solution, as it's the Info structure that you need to change during the life of the application.

So now the question is : "what's the need here"
What we try to resolve is a need for many objets to access to the current selected account.
This one is returned by the call "lrc_->getAccountModel().getAccountInfo(id)" during updateLrc in RingMainWindow.
All listener objects need to access to the const reference returned by this call.
So the easiest way is to record the return of this former call not as a const& but as a const* in RingMainWindow.
Then you pass to listeners a pointer on this pointer: const*&.
Listeners can now obtain the pointed acconout Info at any moments, the same one as recorded by RingMainWindow.

Patch-set: 4
Label: Code-Review=-2