chatview: clearly display banned contacts as such
Whenever the chat view has to display a banned contact, it should:
- Make place call, add to contacts and place audio call buttons
  unclickable
- Hide the send message bar
- Display the banned contact as such in the smart list results (grey bg,
  bold "Banned contact", etc.)

This patch implements this behavior in the GNU/Linux GNOME client.

For the moment the conversation popup is still mentioning invalid
options such as «block contact» or «place call», but since fixing it
requires LRC side changes we will address this issue in a separate
commit.

Change-Id: Ibc8da7d755319acae59a807cc248e3a87bd9d6da
Reviewed-by: Sebastien Blin <sebastien.blin@savoirfairelinux.com>