popupmenu: fix menu for banned accounts
Currently the popup menu stays the same even if the account is banned
which may be very confusing for the user (place call or block contact
options are everything but meaningful in this case !).

In this patch we implement a specific menu for banned accounts.

In addition to that, we also force a chat view refresh when
the filter is updated.

This patch requires LRC side changes in order to work properly.

Change-Id: Iea545fe5cf54520e7aba92c7a87a736bc6d71d27