ringmainwindow: fix view inconsistency bug
Bug: Create two ring accounts A and B and a conversation between them.
The two accounts should not have any more contacts than the other
account, that is A's only contact is B and vice versa. Select the
conversation in A's tab and switch to B. The conversation with A
stays selected _but_ the chat view isn't displayed.

Fix: We should switch to the ring main window before updating the lrc.
Otherwise the conversation is going to be selected by
slotFilterChanged (signal triggered by updateLrc) while the view will
be later changed to welcome view without updating the conversationview
selection.

Change-Id: I2caea45dc2a11d90b409cdb06e9ba9eaa14d9f21