accountcreationwizard: close preview when hidden
Whenever the preview is created and run in the account creation
wizard, there is no way to stop it. This is very annoying because
it grabs the webcam rights without releasing them, and the camera
led stays on, giving the user the impression to be observed.

In this patch we add a mechanism similar to what is done in the
general settings tab: we destroy and create the preview each time
the preview is made visible / hidden. While not the most elegant
solution it allows us to run the preview if and only if it is
displayed & stop it otherwise.

Also, we fix various issues in the account view:
 - when + icon is double clicked, the settings space becomes blank.
   Now if + is clicked while account wizard is open, the account
   wizard gets closed.
 - when an account is selected in the account selection tab while
   the wizard is displayed, nothing "seems to happen". Instead,
   quit the wizard and open the selected tab.
 - when account view is openened for the first time, the settings
   space is blank. Instead, default select an account.

Change-Id: I1e12333e654f70d590886e2aa4f5112154f3068a
Reviewed-by: Sebastien Blin <sebastien.blin@savoirfairelinux.com>