accountcreationwizard: close preview when hidden
Whenever the preview is created and run in the account creation
wizard, there is no way to stop it. This is very annoying because
it grabs the webcam rights without releasing them, and the camera
led stays on, giving the user the impression to be observed.

In this patch we add a mechanism similar to what is done in the
general settings tab: we destroy and create the preview each time
the preview is made visible / hidden. While not the most elegant
solution it allows us to run the preview if and only if it is
displayed & stop it otherwise.

Change-Id: I1e12333e654f70d590886e2aa4f5112154f3068a