chatview: various optimizations and minor bugfixes
Despite of the recent optimization commits, the chatview is still
affected by a significant amount of performance issues. This commit
tries to address some of them:

* Use getElementById instead of querySelector, when possible.
* Use getElementByClassName instead of querySelector, when possible.

querySelector in in average 2x slower than getElementById or
getElementByClassName.

In this commit we also address the following minor bugs:

* printHistoryPart: in the message print loop it doesn't make any
sense to call addOrUpdateMessage if we are at the end of the buffer.
Move the end buffer check at the beginning of the loop. This removes
a js error appearing in the logs.

Change-Id: If6b605868ba6b0b9623ae01c5293064211b58327