chatview: remove loading animations
Animations during conversation loading are heavy and useless. Remove
them. Also remove animations of message status (sending ... and X when
message failed to send) as they are pretty useless and look weird with
lazy loading.

Add a ... icon at the top of the conversation to make it more
intuitive to the user that he can scroll further when a conversation
was lazy loaded.

Add a workaround to stop creating #messages dynamically.

Change-Id: If6183bd58002eb79d459dad5dd3c0b55eecebf86
Gitlab: #836