ringnotify: rewrite notification system
Ring will now use three types of notifications:

1. Call notifications: opened when an incoming call is here
and closed at the end of the call.
2. Request notifications: opened when a new trust request
arrives and closed when the user accepts/refuse/block or
just open the conversation.
3. Chat notifications: arrives with new interactions and if
the current conversation is different or the client not focused. Closed
by Gnome or when the conversation is opened.

Change-Id: I5e5abf20507bac8bb37c429bc929c671fe66bd6b
Gitlab: #868
Reviewed-by: Andreas Traczyk <andreas.traczyk@savoirfairelinux.com>