ringmainwindow: fix avatar issue in settings
Whenever the settings are closed, the avatar manipulation widget is
destroyed. This is because we want to make sure the video preview is
always stopped properly. However the avatar widget is not reinitialized
when going back to the settings, leading to a missing avatar widget.

Add missing call to new_account_settings_view_show in enterSettingsView().

Change-Id: If3b97a92248cba43173a666833a1b372f1d19f4e
Reviewed-by: Sebastien Blin <sebastien.blin@savoirfairelinux.com>