Update patch set 36
Patch Set 36: Code-Review-1

(2 comments)

Ok. Just checked. Two typos, nothing to say on the code itself. Nice!

I only got one big issue. The busy tone is launched when I click on "record an audio message" and I can't stop it. It's not related to the code of this patch, but still, I prefer to not push this still the other issue is fixed.

Patch-set: 36