mainwindow: alert user if minimizing in systray with calls
When the user click on close when systray is enabled, the app will
only be minimized, not closed. So the call will continue to turn in
background. This means that the camera will be still visible to the
peer. This patches had a dialog to warn the user about that.

Change-Id: Ib5afe0415156aa5d7d29ee66acec89865e59e73d
Gitlab: #1000