accountcreationwizard: disable Jami account encryption by default
Move the password and password confirmation entries behind a switch
that is disabled by default.  Upon enabling the switch, the entries
will be shown, along with an explanation of what the user is about to
do and that they will not be able to recover their account if they
lose their password.

GitLab: #1197
Change-Id: I6aad54e502f6ed9ba0176217a5c090fcdb2c4db1