pixmaps: add and use Jami logo with white text with dark theme
accountcreationwizard, dialogs, mainwindow, welcomeview: use the Jami
logo with the white text when using a dark theme.

utils/drawing: bring over the logic for detection of dark theme from
'mainwindow' to here, so that it could also be used by 'dialogs' for
the about dialog.

dialogs: change from .c to .cpp so we can include utils/drawing;
remove some unused code; and add my name to the authors list.

GitLab: #1221
Change-Id: If466427d7fc63bbd1b0fbd739e5b82592001f547