conversationsview: restore merging ongoing calls with drag-and-drop
I removed this in an earlier commit, but turns out this is something
people depend on, and to my knowledge there is no other way to do this
for two ongoing calls at the moment.  Perhaps remove later only if we
have another (better) way of doing the same action.

Change-Id: Ic5f67a9ed59e870ad96a7f02eed87ea529c588ad