avatarmanipulation: fix recent regression in avatar preview
The recent 990ffe0077a495163f9641854d3cbf94f36dfe99 commit did not
make changes needed for avatarmanipulation due to recent LRC change
in video threading.  This fixes the regression, and makes the avatar
preview behave similarly to the 'Settings > Media' video preview, and
reuses an existing renderer when possible (e.g. when in an ongoing
call).

Change-Id: I1dfa386318acd7178c1840b79fa173c85128f901