chatview: add insert emoji button
Also enable local storage for recent emoji support.  As mentioned in
the commit message of the corresponding change in lrc, we resort to
the ugly hack of embedding the two set of CSS styles needed by the
picker, which should be injected/loaded into the DOM by emoji.js, but
for some reason is not, by client-gnome's webkit view.

Change-Id: Id34ce5131e8027a550a8443aa518dd53e4fa9176