version: Remove the build date and unify version string.
Users want to know what sources their software was built from, not the
date it was built at.  Also make the version string a single thing,
which makes it easier to handle the following two scenarios:

1. Development build.  The version is set to the git hash, which
should communicate this clearly enough.

2. Release build.  The version is whatever we choose to be, currently
something like "20210606.1.e2f9490", which already includes
the (super-project) git hash information.

* CMakeLists.txt (CLIENT_BUILD_DATE): Remove variable.
(PROJECT_VERSION): Set it to either the manually specified string or
to the Git hash.  Print a warning in case it is not found.
(git_revision): Remove target.
* src/dialogs.cpp (about_dialog): Adjust the displayed version
information.
* src/client_options.c (option_version_cb): Likewise.
* src/client.cpp (client_startup): Only print VERSION; remove
CLIENT_REVISION.
* cmake/GitRevision.cmake: Likewise.
* src/revision.h.in: Delete file.

Change-Id: If893b5781ddf7acf54dabdaab37f27f8b87a9ea1